refactor: move ModifierExtensions to modifier package
